+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/Scripts.hs
@@ -20,7 +20,7 @@
 {-# OPTIONS_GHC -Wno-orphans #-}
 
 module Cardano.Ledger.Shelley.Scripts (
-  MultiSig,
+  MultiSig (..),
   ShelleyEraScript (..),
   pattern RequireSignature,
   pattern RequireAllOf,
@@ -30,12 +30,11 @@ module Cardano.Ledger.Shelley.Scripts (
   validateMultiSig,
   nativeMultiSigTag,
   eqMultiSigRaw,
-  MultiSigRaw,
+  MultiSigRaw (..),
 )
 where
 
 import Cardano.Ledger.Binary (
-  Annotator (..),
   DecCBOR (decCBOR),
   EncCBOR (..),
   ToCBOR,
@@ -50,7 +49,6 @@ import Cardano.Ledger.Core
 import Cardano.Ledger.Keys.WitVKey (witVKeyHash)
 import Cardano.Ledger.MemoBytes (
   EqRaw (..),
-  Mem,
   MemoBytes,
   Memoized (..),
   decodeMemoized,
@@ -160,11 +158,6 @@ instance ShelleyEraScript ShelleyEra where
 
 deriving newtype instance NFData (MultiSig era)
 
-deriving via
-  Mem (MultiSigRaw era)
-  instance
-    Era era => DecCBOR (Annotator (MultiSig era))
-
 instance EqRaw (MultiSig era) where
   eqRaw = eqMultiSigRaw
 
@@ -202,22 +195,6 @@ instance Era era => DecCBOR (MultiSigRaw era) where
       1 -> (,) 2 . MultiSigAllOf <$> decCBOR
       2 -> (,) 2 . MultiSigAnyOf <$> decCBOR
       3 -> (,) 3 <$> (MultiSigMOf <$> decCBOR <*> decCBOR)
-      k -> invalidKey k
-
-instance Era era => DecCBOR (Annotator (MultiSigRaw era)) where
-  decCBOR = decodeRecordSum "MultiSig" $
-    \case
-      0 -> (,) 2 . pure . MultiSigSignature . KeyHash <$> decCBOR
-      1 -> do
-        multiSigs <- sequence <$> decCBOR
-        pure (2, MultiSigAllOf <$> multiSigs)
-      2 -> do
-        multiSigs <- sequence <$> decCBOR
-        pure (2, MultiSigAnyOf <$> multiSigs)
-      3 -> do
-        m <- decCBOR
-        multiSigs <- sequence <$> decCBOR
-        pure (3, MultiSigMOf m <$> multiSigs)
       k -> invalidKey k
 
 -- | Check the equality of two underlying types, while ignoring their binary