+++ libs/cardano-ledger-core/src/Cardano/Ledger/Core.hs
@@ -73,7 +73,6 @@ import Cardano.Ledger.Address (
  )
 import Cardano.Ledger.BaseTypes (ProtVer (..))
 import Cardano.Ledger.Binary (
-  Annotator,
   DecCBOR,
   DecShareCBOR (Share),
   EncCBOR,
@@ -126,7 +125,7 @@ class
   , EraPParams era
   , -- NFData (Tx era), TODO: Add NFData constraints to Crypto class
     NoThunks (Tx era)
-  , DecCBOR (Annotator (Tx era))
+  , DecCBOR (Tx era)
   , EncCBOR (Tx era)
   , ToCBOR (Tx era)
   , Show (Tx era)
@@ -185,7 +184,7 @@ class
   , EraTxCert era
   , EraPParams era
   , HashAnnotated (TxBody era) EraIndependentTxBody
-  , DecCBOR (Annotator (TxBody era))
+  , DecCBOR (TxBody era)
   , EncCBOR (TxBody era)
   , ToCBOR (TxBody era)
   , NoThunks (TxBody era)
@@ -453,7 +452,7 @@ class
   , NoThunks (TxAuxData era)
   , ToCBOR (TxAuxData era)
   , EncCBOR (TxAuxData era)
-  , DecCBOR (Annotator (TxAuxData era))
+  , DecCBOR (TxAuxData era)
   , HashAnnotated (TxAuxData era) EraIndependentTxAuxData
   ) =>
   EraTxAuxData era
@@ -487,7 +486,7 @@ class
   , NoThunks (TxWits era)
   , ToCBOR (TxWits era)
   , EncCBOR (TxWits era)
-  , DecCBOR (Annotator (TxWits era))
+  , DecCBOR (TxWits era)
   ) =>
   EraTxWits era
   where
@@ -510,7 +509,7 @@ hashScriptTxWitsL ::
   Lens (TxWits era) (TxWits era) (Map ScriptHash (Script era)) [Script era]
 hashScriptTxWitsL =
   lens
-    (\wits -> wits ^. scriptTxWitsL)
+    (^. scriptTxWitsL)
     (\wits ss -> wits & scriptTxWitsL .~ Map.fromList [(hashScript s, s) | s <- ss])
 {-# INLINEABLE hashScriptTxWitsL #-}
 
@@ -541,7 +540,7 @@ class
   , EqRaw (Script era)
   , ToCBOR (Script era)
   , EncCBOR (Script era)
-  , DecCBOR (Annotator (Script era))
+  , DecCBOR (Script era)
   , NoThunks (Script era)
   , SafeToHash (Script era)
   , Eq (NativeScript era)
@@ -549,7 +548,7 @@ class
   , NFData (NativeScript era)
   , NoThunks (NativeScript era)
   , EncCBOR (NativeScript era)
-  , DecCBOR (Annotator (NativeScript era))
+  , DecCBOR (NativeScript era)
   ) =>
   EraScript era
   where
@@ -609,7 +608,7 @@ class
   , Eq (TxSeq era)
   , Show (TxSeq era)
   , EncCBORGroup (TxSeq era)
-  , DecCBOR (Annotator (TxSeq era))
+  , DecCBOR (TxSeq era)
   ) =>
   EraSegWits era
   where