+++ libs/cardano-ledger-core/src/Cardano/Ledger/Keys/WitVKey.hs
@@ -20,11 +20,7 @@ import Cardano.Crypto.DSIGN.Class (
   decodeSignedDSIGN,
   encodeSignedDSIGN,
  )
-import Cardano.Ledger.Binary (
-  Annotator (..),
-  DecCBOR (..),
-  EncCBOR (..),
- )
+import Cardano.Ledger.Binary (DecCBOR (..), EncCBOR (..))
 import qualified Cardano.Ledger.Binary.Plain as Plain
 import Cardano.Ledger.Hashes (
   EraIndependentTxBody,
@@ -87,10 +83,6 @@ instance Typeable kr => Plain.FromCBOR (WitVKey kr) where
       WitVKey <$> Plain.fromCBOR <*> decodeSignedDSIGN
 
 instance Typeable kr => EncCBOR (WitVKey kr)
-
-instance Typeable kr => DecCBOR (Annotator (WitVKey kr)) where
-  decCBOR = pure <$> decCBOR
-  {-# INLINE decCBOR #-}
 
 instance Typeable kr => DecCBOR (WitVKey kr)