+++ libs/cardano-ledger-core/src/Cardano/Ledger/MemoBytes/Internal.hs
@@ -31,7 +31,6 @@
 module Cardano.Ledger.MemoBytes.Internal (
   MemoBytes (.., Memo),
   MemoHashIndex,
-  Mem,
   mkMemoBytes,
   mkMemoBytesStrict,
   getMemoBytesType,
@@ -71,15 +70,13 @@ where
 import Cardano.Crypto.Hash (HashAlgorithm (hashAlgorithmName))
 import Cardano.Ledger.Binary (
   Annotated (..),
-  Annotator (..),
   DecCBOR (decCBOR),
   Decoder,
   EncCBOR,
   Version,
   decodeAnnotated,
-  decodeFullAnnotator,
+  decodeFull',
   serialize,
-  withSlice,
  )
 import Cardano.Ledger.Binary.Coders (Encode, encode, runE)
 import qualified Cardano.Ledger.Binary.Plain as Plain
@@ -94,7 +91,6 @@ import qualified Data.ByteString.Short as SBS (length)
 import Data.Coerce
 import Data.MemPack
 import Data.MemPack.Buffer (Buffer)
-import qualified Data.Text as T
 import Data.Typeable
 import GHC.Base (Type)
 import GHC.Generics (Generic)
@@ -131,26 +127,11 @@ byteCountMemoBytes = packedByteCount . mbBytes
 packMemoBytesM :: MemoBytes t -> Pack s ()
 packMemoBytesM = packM . mbBytes
 
-unpackMemoBytesM ::
-  forall t b.
-  (Typeable t, DecCBOR (Annotator t), Buffer b) =>
-  Version ->
-  Unpack b (MemoBytes t)
+unpackMemoBytesM :: (DecCBOR t, Buffer b) => Version -> Unpack b (MemoBytes t)
 unpackMemoBytesM v = unpackM >>= decodeMemoBytes v
 
-decodeMemoBytes ::
-  forall t m.
-  (Typeable t, DecCBOR (Annotator t), MonadFail m) =>
-  Version ->
-  ByteString ->
-  m (MemoBytes t)
-decodeMemoBytes v bs =
-  either (fail . show) pure $
-    decodeFullAnnotator
-      v
-      (T.pack (show (typeRep (Proxy @t))))
-      decCBOR
-      (BSL.fromStrict bs)
+decodeMemoBytes :: (DecCBOR t, MonadFail m) => Version -> ByteString -> m (MemoBytes t)
+decodeMemoBytes v = either (fail . show) pure . decodeFull' v
 
 type family MemoHashIndex (t :: Type) :: Type
 
@@ -158,14 +139,6 @@ deriving instance NFData t => NFData (MemoBytes t)
 
 instance Typeable t => Plain.ToCBOR (MemoBytes t) where
   toCBOR (MemoBytes _ bytes _hash) = Plain.encodePreEncoded (fromShort bytes)
-
-instance
-  (Typeable t, DecCBOR (Annotator t)) =>
-  DecCBOR (Annotator (MemoBytes t))
-  where
-  decCBOR = do
-    (Annotator getT, Annotator getBytes) <- withSlice decCBOR
-    pure (Annotator (\fullbytes -> mkMemoBytes (getT fullbytes) (getBytes fullbytes)))
 
 instance DecCBOR t => DecCBOR (MemoBytes t) where
   decCBOR = decodeMemoized decCBOR
@@ -192,10 +165,6 @@ shorten :: BSL.ByteString -> ShortByteString
 shorten x = toShort (toStrict x)
 {-# DEPRECATED shorten "As unused. Use `toShort` `.` `toStrict` instead" #-}
 
--- | Useful when deriving DecCBOR(Annotator T)
--- deriving via (Mem T) instance DecCBOR (Annotator T)
-type Mem t = Annotator (MemoBytes t)
-
 -- | Constructor that takes the underlying type and the original bytes as lazy
 -- `BSL.ByteString`.
 --
@@ -207,10 +176,8 @@ mkMemoBytes t = mkMemoBytesStrict t . toStrict
 -- | Same as `mkMemoBytes`, but with strict bytes
 mkMemoBytesStrict :: forall t. t -> ByteString -> MemoBytes t
 mkMemoBytesStrict t bs =
-  MemoBytes
-    t
-    (toShort bs)
-    (makeHashWithExplicitProxys (Proxy @(MemoHashIndex t)) bs)
+  MemoBytes t (toShort bs) $
+    makeHashWithExplicitProxys (Proxy @(MemoHashIndex t)) bs
 
 -- | Turn a MemoBytes into a string, showing both its internal structure and its original bytes.
 --   Useful since the Show instance of MemoBytes does not display the original bytes.