+++ libs/cardano-protocol-tpraos/src/Cardano/Protocol/TPraos/BHeader.hs
@@ -19,7 +19,8 @@
 module Cardano.Protocol.TPraos.BHeader (
   HashHeader (..),
   PrevHash (..),
-  BHeader (BHeader),
+  BHeader (BHeader, ..),
+  BHeaderRaw (..),
   BHBody (..),
   LastAppliedBlock (..),
   BoundedNatural (bvValue, bvMaxValue),
@@ -60,7 +61,6 @@ import Cardano.Ledger.BaseTypes (
  )
 import Cardano.Ledger.BaseTypes.NonZero (nonZero, (%.))
 import Cardano.Ledger.Binary (
-  Annotator (..),
   Case (..),
   DecCBOR (decCBOR),
   DecCBORGroup (..),
@@ -96,7 +96,6 @@ import Cardano.Ledger.Hashes (
  )
 import Cardano.Ledger.Keys (VKey)
 import Cardano.Ledger.MemoBytes (
-  Mem,
   MemoBytes,
   MemoHashIndex,
   Memoized (..),
@@ -274,9 +273,6 @@ instance Crypto c => DecCBOR (BHeaderRaw c) where
     decodeRecordNamed "BHeaderRaw" (const 2) $
       BHeaderRaw <$> decCBOR <*> decodeSignedKES
 
-instance Crypto c => DecCBOR (Annotator (BHeaderRaw c)) where
-  decCBOR = pure <$> decCBOR
-
 newtype BHeader c = BHeaderConstr (MemoBytes (BHeaderRaw c))
   deriving (Generic)
   deriving newtype (Eq, Show, NoThunks, Plain.ToCBOR, DecCBOR, SafeToHash)
@@ -289,11 +285,6 @@ instance Crypto c => EncCBOR (BHeader c)
 
 instance Memoized (BHeader c) where
   type RawType (BHeader c) = BHeaderRaw c
-
-deriving via
-  Mem (BHeaderRaw c)
-  instance
-    Crypto c => DecCBOR (Annotator (BHeader c))
 
 pattern BHeader ::
   Crypto c =>