+++ libs/cardano-ledger-core/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 1.18.0.0
 
+* Add `eraDecoderWithBytes`
+* Move `Annotator` instances to `testlib`
+* Expose `MkData` constructor.
+* Remove `Text` parameter from `translateEraThroughCBOR`
 * Add `keyHashWitnessesTxWits`
 * Remove export of `mkMemoBytes` from `Cardano.Ledger.MemoBytes` as unsafe
 * Add `CanGetChainAccountState` and `CanSetChainAccountState`